+++ esp-metadata/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Clock tree modeling (#4501)
 - Encode pin restrictions in the generated data (#4790)
 - Initial ESP32-C5 support (#4859, #4866)
+- DMA properties and peripheral selector values (#4944)
 
 ### Changed